api/pyproject.toml
[project]
name = "api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "daphne>=4.2.1",
    "django>=6.0",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.1",
    "dotenv>=0.9.9",
    "psycopg",
    "redis>=7.1.0",
    "requests>=2.32.5",
]
